+++ CHANGELOG.md
@@ -9,12 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Disable `system.copySystemConfiguration` to support flakes.
+
 ## 2.1.0 - 2026-03-10
 
 ### Added
 
 - Add `programs.hyprland.plugins.hyprexpo.enable`
   to make Hyprexpo plugin could be disabled.
+
 - Add `programs.hyprland.plugins.hyprexpo.enable`
   to make it could be not installed
   in case that Hyprexpo and Hyprland do not share the same version